+++ src/main/java/org/fairdatapoint/vocabulary/R3D.java
@@ -1,0 +1,38 @@
+/**
+ * The MIT License
+ * Copyright © 2016-2024 FAIR Data Team
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package org.fairdatapoint.vocabulary;
+
+import org.eclipse.rdf4j.model.IRI;
+
+import static org.fairdatapoint.util.ValueFactoryHelper.i;
+
+public final class R3D {
+
+    public static final String NAMESPACE = "http://www.re3data.org/schema/3-0#";
+
+    public static final IRI COUNTRY = i(NAMESPACE + "country");
+    public static final IRI DATACATALOG = i(NAMESPACE + "dataCatalog");
+    public static final IRI REPOSITORY = i(NAMESPACE + "Repository");
+    public static final IRI REPOSITORYIDENTIFIER = i(NAMESPACE + "repositoryIdentifier");
+
+}